Extract rotation stuff into separate class
